+++ src/internal/quota/providers/copilot.go
@@ -103,7 +103,7 @@ func copilotWindow(s *CopilotQuotaSnapshot) quota.Quota {
 		if used < 0 {
 			used = 0
 		}
-		return quota.Quota{Used: used, Limit: s.Entitlement, Unit: "count"}
+		return quota.Quota{Used: float64(used), Limit: float64(s.Entitlement), Unit: "count"}
 	}
 	if s.PercentRemaining != nil {
 		used := 100 - int(*s.PercentRemaining+0.5)
@@ -113,10 +113,10 @@ func copilotWindow(s *CopilotQuotaSnapshot) quota.Quota {
 		if used > 100 {
 			used = 100
 		}
-		return quota.Quota{Used: used, Limit: 100, Unit: "percentage"}
+		return quota.Quota{Used: float64(used), Limit: 100, Unit: "percentage"}
 	}
 	if s.OverQuotaUsedPercent != nil {
-		return quota.Quota{Used: clampPercent(*s.OverQuotaUsedPercent), Limit: 100, Unit: "percentage"}
+		return quota.Quota{Used: float64(clampPercent(*s.OverQuotaUsedPercent)), Limit: 100, Unit: "percentage"}
 	}
 	return quota.Quota{Used: 0, Limit: 100, Unit: "percentage"}
 }
